+++ Kafkaesque.Tests/TestLongHistory.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -35,19 +36,25 @@ namespace Kafkaesque.Tests
             var directoryInfo = new DirectoryInfo(logDirectoryPath);
             var logDirectory = new LogDirectory(directoryInfo);
 
+            var writeStopwatch = Stopwatch.StartNew();
+
             // write everything
             var writer = logDirectory.GetWriter();
             Using(writer);
             await writer.WriteManyAsync(messages.Select(Encoding.UTF8.GetBytes));
 
+            var elapsedSecondsWriting = writeStopwatch.Elapsed.TotalSeconds;
+
+            Console.WriteLine($"Wrote {count} messages in {elapsedSecondsWriting:0.0} s - that's {count/elapsedSecondsWriting:0.0} msg/s");
+
             directoryInfo.DumpDirectoryContentsToConsole();
 
             // read it back
             var reader = logDirectory.GetReader();
-
+            var readStopwatch = Stopwatch.StartNew();
             var expectedMessageNumber = 0;
 
-            foreach (var message in reader.Read(cancellationToken: CancelAfter(TimeSpan.FromSeconds(10))).Take(count))
+            foreach (var message in reader.Read(cancellationToken: CancelAfter(TimeSpan.FromSeconds(20))).Take(count))
             {
                 var text = Encoding.UTF8.GetString(message.Data);
                 var parts = text.Split('/');
@@ -80,6 +87,9 @@ namespace Kafkaesque.Tests
 
                 expectedMessageNumber++;
             }
+
+            var elapsedSeconds = readStopwatch.Elapsed.TotalSeconds;
+            Console.WriteLine($"Read {count} messages in {elapsedSeconds:0.0} s - that's {count/elapsedSeconds:0.0} msg/s");
 
             Assert.That(expectedMessageNumber, Is.EqualTo(count));
         }
@@ -133,10 +143,10 @@ namespace Kafkaesque.Tests
             });
 
             var reader = logDirectory.GetReader();
-
             var expectedMessageNumber = 0;
+            var stopwatch = Stopwatch.StartNew();
 
-            foreach (var message in reader.Read(cancellationToken: CancelAfter(TimeSpan.FromSeconds(10))).Take(count))
+            foreach (var message in reader.Read(cancellationToken: CancelAfter(TimeSpan.FromSeconds(20))).Take(count))
             {
                 var text = Encoding.UTF8.GetString(message.Data);
                 var parts = text.Split('/');
@@ -169,6 +179,9 @@ namespace Kafkaesque.Tests
 
                 expectedMessageNumber++;
             }
+
+            var elapsedSeconds = stopwatch.Elapsed.TotalSeconds;
+            Console.WriteLine($"Read {count} messages in {elapsedSeconds:0.0} s - that's {count/elapsedSeconds:0.0} msg/s");
 
             Assert.That(expectedMessageNumber, Is.EqualTo(count));
         }