+++ Kafkaesque/LogDirectory.cs
@@ -11,20 +11,24 @@ namespace Kafkaesque
     public class LogDirectory
     {
         readonly string _directoryPath;
+        readonly Settings _settings;
 
         /// <summary>
         /// Creates a log directory, using the directory specified by <paramref name="directoryInfo"/> to store the files.
         /// </summary>
-        public LogDirectory(DirectoryInfo directoryInfo) : this(directoryInfo?.FullName)
+        public LogDirectory(DirectoryInfo directoryInfo, Settings settings = null) : this(directoryInfo?.FullName, settings)
         {
         }
 
         /// <summary>
         /// Creates a log directory, using the directory specified by <paramref name="directoryPath"/> to store the files.
         /// </summary>
-        public LogDirectory(string directoryPath)
+        public LogDirectory(string directoryPath, Settings settings = null)
         {
             _directoryPath = directoryPath ?? throw new ArgumentNullException(nameof(directoryPath), "Please pass a directory path to the log directory");
+            _settings = settings ?? new Settings();
+
+            _settings.Validate();
 
             if (!Directory.Exists(directoryPath))
             {
@@ -38,12 +42,12 @@ namespace Kafkaesque
         /// Optionally pass a <paramref name="cancellationToken"/> if you want to be able to abort the acquisition attempt
         /// prematurely.
         /// </summary>
-        public LogWriter GetWriter(CancellationToken cancellationToken = default) => new LogWriter(_directoryPath, cancellationToken);
+        public LogWriter GetWriter(CancellationToken cancellationToken = default) => new LogWriter(_directoryPath, cancellationToken, _settings);
 
         /// <summary>
         /// Gets a log reader.
         /// </summary>
-        public LogReader GetReader() => new LogReader(_directoryPath);
+        public LogReader GetReader() => new LogReader(_directoryPath, _settings);
 
         static void CreateLogDirectory(string directoryPath)
         {