+++ CHANGELOG.md
@@ -4,3 +4,11 @@
 
 * Boom! Initial version.
 
+## 0.0.5
+
+* Ability to configure a few things
+* Can delete files now
+
+## 0.0.6
+
+* Remove Serilog dependency